refactor: split scheduled learning reducer
