+++ experiments/scheduled-task-learning/benchmark-core/benchmark_learning/learning_replay/__init__.py
@@ -1,0 +1,6 @@
+"""Pure scheduled-learning reducer over the closed ordered replay event log."""
+
+from .reducer import replay
+from .state import initial_state
+
+__all__ = ["initial_state", "replay"]
